pyproject.toml
[tool.poetry]
name = "strawberry_django_project"
version = "0.1.0"
description = ""
authors = ["ronald.williams <ronald.williams@onepeloton.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
strawberry-graphql = "^0.124.0"
strawberry-graphql-django = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
